Shiny bindings for HARviewer

Shiny bindings for HARviewer

HARviewerOutput(outputId, width = "100%", height = "400px")

renderHARviewer(expr, env = parent.frame(), quoted = FALSE)

Arguments

outputId

output variable to read from

width, height

Must be a valid CSS unit (like '100%', '400px', 'auto') or a number, which will be coerced to a string and have 'px' appended. height will probably not have an effect; instead, use the height parameter in HARviewer.

expr

An expression that generates a HARviewer

env

The environment in which to evaluate expr.

quoted

Is expr a quoted expression (with quote())? This is useful if you want to save an expression in a variable.

Examples

## Not run: ------------------------------------
# library(shiny)
# library(HARtools)
# har <- readHAR(system.file(package = "HARtools", "exdata",
#                            "r-project.org.161028_W2_11MA.har"))
# hv <- HARviewer(har)
# 
# shinyApp(
#   ui = fluidPage(
#     HARviewerOutput("myHAR")
#   ),
#   server = function(input, output) {
#     output$myHAR <- renderHARviewer(hv)
#   }
# )
## ---------------------------------------------